feat: add graph2plantuml code logic. chore: refactors logic to have a clear split of interfaces: input -> graph -> svg -> output.
